refactor: Split diversion rules into smaller modules
Put rule conditions and actions into their into module

Related #2378
